backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.11",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "redis>=7.2.0",
]
